+++ src/ext/conf/BrowserDetect.js
@@ -1,4 +1,7 @@
-import Debug from './Debug.js';
+if (process.env.CHANNEL !== 'stable') {
+  console.log('Loaded BrowserDetect');
+}
+
 
 const BrowserDetect = {
   firefox: process.env.BROWSER === 'firefox',
@@ -7,7 +10,7 @@ const BrowserDetect = {
   processEnvBrowser: process.env.BROWSER,
 } 
 
-if (Debug.debug) {
+if (process.env.CHANNEL !== 'stable') {
   console.log("Loading: BrowserDetect.js\n\nprocess.env.BROWSER:", process.env.BROWSER, "Exporting BrowserDetect:", BrowserDetect);
 }