+++ src/ext/lib/comms/Comms.js
@@ -1,6 +1,10 @@
 import Debug from '../../conf/Debug';
 import BrowserDetect from '../../conf/BrowserDetect';
 
+if (process.env.CHANNEL !== 'stable'){
+  console.log("Loading Comms");
+}
+
 class Comms {
   static async sendMessage(message){
 
@@ -30,6 +34,10 @@ class Comms {
     }
   }
 
+}
+
+if (process.env.CHANNEL !== 'stable'){
+  console.log("Comms loaded");
 }
 
 export default Comms;